Changing helper for handlebars function
