+++ src/routes/users.js
@@ -5,4 +5,13 @@ const api = express.Router()
 
 api.get('/pruebas', UsersController.pruebas)
 
+api.post('/login', (req, res) => {
+    const isRegistered = users.find((user) => user.id === req.body.idNumber.toString() && user.password === req.body.password.toString())
+    if (isRegistered) {
+        res.redirect('/courses')
+    } else {
+        res.redirect(401, '/')
+    }
+})
+
 module.exports = api
