+++ src/routes/views.js
@@ -1,4 +1,5 @@
 const express = require('express')
+const courses = require('../../courses.json')
 
 const app = express()
 
@@ -6,7 +7,8 @@ app.get('/', (req, res) => {
     res.render('index')
 })
 app.get('/courses', (req, res) => {
-    res.render('courses')
+    console.log(courses)
+    res.render('courses', { courses: courses.courses })
 })
 app.get('/register', (req, res) => {
     res.render('register')
