+++ views/courses.hbs
@@ -1,52 +1,73 @@
 {{#> page_wrapper }}
-  <h2>Listado de cursos</h2>
+<h2>Listado de cursos</h2>
 
-
-  <h4>Agregar Curso</h4>
-  <div class="row">
-    <form class="col s12" action="/api/pruebas" method="GET">
-      <div class="row">
-        <div class="input-field col s6">
-          <input id="id" type="text" class="validate" required>
-          <label for="id">Id</label>
-        </div>
-        <div class="input-field col s6">
-          <input id="Name" type="text" class="validate" required>
-          <label for="Name">Nombre del curso</label>
-        </div>
+<h4>Agregar Curso</h4>
+<div class="row">
+  <form class="col s12" action="/api/pruebas" method="GET">
+    <div class="row">
+      <div class="input-field col s6">
+        <input id="id" type="text" class="validate" required>
+        <label for="id">Id</label>
       </div>
-      <div class="row">
-        <div class="input-field col s6">
-          <input id="Modality" type="text" class="validate" required>
-          <label for="Modality">Modalidad</label>
-        </div>
-        <div class="input-field col s6">
-          <input id="Intensity" type="text" class="validate" required>
-          <label for="Intensity">Intensidad horaria</label>
-        </div>
+      <div class="input-field col s6">
+        <input id="Name" type="text" class="validate" required>
+        <label for="Name">Nombre del curso</label>
       </div>
-      <div class="row">
-        <div class="input-field col s6">
-          <input id="price" type="text" class="validate" required>
-          <label for="price">Valor</label>
-        </div>
-        <div class="input-field col s6">
-          <input id="Description" type="text" class="validate materialize-textarea" required>
-          <label for="Description">Descripción</label>
-        </div>
+    </div>
+    <div class="row">
+      <div class="input-field col s6">
+        <input id="Modality" type="text" class="validate" required>
+        <label for="Modality">Modalidad</label>
       </div>
-      <button class="btn waves-effect waves-light" type="submit">Enviar
-        <i class="material-icons right">send</i>
-      </button>
-    </form>
+      <div class="input-field col s6">
+        <input id="Intensity" type="text" class="validate" required>
+        <label for="Intensity">Intensidad horaria</label>
+      </div>
+    </div>
+    <div class="row">
+      <div class="input-field col s6">
+        <input id="price" type="text" class="validate" required>
+        <label for="price">Valor</label>
+      </div>
+      <div class="input-field col s6">
+        <input id="Description" type="text" class="validate materialize-textarea" required>
+        <label for="Description">Descripción</label>
+      </div>
+    </div>
+    <button class="btn waves-effect waves-light" type="submit">Enviar
+      <i class="material-icons right">send</i>
+    </button>
+  </form>
 
-  </div>
+</div>
 
-  <hr>
+<hr>
 
-  <h4>Cursos</h4>
+<h4>Cursos</h4>
 
-      {{{listar}}}
+<table>
+  <thead>
+    <tr>
+      <th>Id</th>
+      <th>Nombre</th>
+      <th>Modalidad</th>
+      <th>Intensidad horaria</th>
+      <th>Valor</th>
+      <th>Descripción</th>
+    </tr>
+  </thead>
+  <tbody>
+    {{#each courses}}
+      <tr>
+        <td>{{this.id}}</td>
+        <td>{{this.name}}</td>
+        <td>{{this.mode}}</td>
+        <td>{{this.intensity}} hr</td>
+        <td>${{this.price}}</td>
+        <td>{{this.description}}</td>
+      </tr>
+    {{/each}}
+  </tbody>
+</table>
 
-
-  {{/ page_wrapper }}
+{{/ page_wrapper }}
